+++ src/app/model/types.ts
@@ -1,0 +1,8 @@
+export type Status = 'idle' | 'loading' | 'succeeded' | 'failed';
+export type AppError = string | null;
+
+export type AppInitialState = {
+    status: Status;
+    error: AppError;
+    isInitialized: boolean;
+};
