+++ src/features/auth/model/types.ts
@@ -1,0 +1,6 @@
+import type { CustomerSignInResult } from '@commercetools/platform-sdk';
+
+export type AuthState = {
+    user: CustomerSignInResult | null;
+    isLoggedIn: boolean;
+};
